requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
googlemaps==4.10.0
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
httpx==0.25.2
unidecode==1.3.7
